requirements.txt
# Core
tornado==2.4.1
sqlalchemy==0.8.0
pyyaml
colander>=0.9.9
pypubsub
# tornado-redis==0.0.7
pika==0.9.5
requests==1.1.0
python-dateutil>=1.5
tornadio2==0.0.4

# For database migrations.
alembic

# MySQL service.
torndb
MySQL-python

# Postgres service
momoko==0.5.0

# MongoDB (should be optional, but isn't yet)
pymongo

# For AWS related services.
boto

# Testing.
mock
coverage

# For documentation.
sphinx==1.1.3
